refactor: rename query paths and remove unused files
- Updated import paths from '@queries' to '@graph' for consistency across the codebase.
- Removed obsolete query files related to currencies and rodeo tokens to eliminate duplication and improve organization.
- Adjusted related components to reflect the new import paths, ensuring functionality remains intact.
