+++ src/graph/currencies/currencies.graph.ts
