+++ src/lib/hooks/queries/use-rodeo-tokens.query.ts
@@ -1,8 +1,8 @@
 import { useEffect, useMemo } from 'react';
 import { useQuery } from '@apollo/client';
+import { AMBRODEO_TOKENS } from '@graph/rodeo-tokens';
+import { RodeoToken } from '@graph/rodeo-tokens/types';
 import { ApolloEndpointsKeys } from '@lib/apollo';
-import { AMBRODEO_TOKENS } from '@queries/rodeo-tokens';
-import { RodeoToken } from '@queries/rodeo-tokens/types';
 import { useRodeoTokensStore } from '@store/rodeo-tokens';
 
 type TokensQuery = {
