+++ tsconfig.json
@@ -27,7 +27,7 @@
       "@utils/*": ["src/utils/*"],
       "@lib/*": ["src/lib/*"],
       "@styles/*": ["src/styles/*"],
-      "@queries/*": ["src/queries/*"],
+      "@graph/*": ["src/graph/*"],
       "@store/*": ["src/store/*"]
     }
   }
